backend/requirements_dev.txt
black==20.8b1
django-debug-toolbar==3.2.1
django-extensions==3.1.1
django-read-only==1.3.0
django-seed==0.2.2
django-stubs==1.7.0
factory-boy==3.2.0
ipython==7.21.0
mypy==0.812
openpyxl==3.0.7
pandas==1.2.4
pdbpp==0.10.2
Pygments==2.8.1
git+https://github.com/phillipdupuis/pydantic-to-typescript.git@master  # we need https://github.com/phillipdupuis/pydantic-to-typescript/pull/4
pygraphviz==1.7
pytest-django==4.1.0
pytest==6.2.2
python-language-server==0.36.2
typing-extensions==3.7.4.3
